Struct cyclone::core::vector::Vector3

pub struct Vector3 {
    pub x: Real,
    pub y: Real,
    pub z: Real,
    pub pad: Real,
}

Struct representing 3-Dimensional vectors

Fields
x: Real y: Real z: Real pad: Real
Methods

impl Vector3

Implement Vector3

pub fn new(x: Real, y: Real, z: Real) -> Self

pub fn invert(self) -> Self

Takes ownership of self Returns new instance of self that is inverted

pub fn invert_inplace(&mut self)

Mutates self Inverts self without returning

pub fn invert_clone(&self) -> Self

Does not take ownership of self Returns a new instance that is inverted from self

pub fn magnitude(&self) -> Real

pub fn square_magnitude(&self) -> Real

pub fn normalize(self) -> Self

pub fn normalize_inplace(&mut self)

pub fn normalize_clone(&self) -> Self

pub fn cross(self, rhs: Self) -> Self

pub fn cross_clone(&self, rhs: &Self) -> Self

pub fn dot(self, rhs: Self) -> Real

pub fn create_orthonormal_basis(&self, b: &Self) -> Option<(Self, Self)>

pub fn find_angle(&self, b: &Self) -> Real